refactor eoa store
