+++ executors/src/eoa/mod.rs
@@ -1,6 +1,7 @@
 pub mod error_classifier;
 pub mod store;
 pub mod worker;
+
 pub use error_classifier::{EoaErrorMapper, EoaExecutionError, RecoveryStrategy};
 pub use store::{EoaExecutorStore, EoaTransactionRequest};
 pub use worker::{EoaExecutorWorker, EoaExecutorWorkerJobData};
