+++ server/src/execution_router/mod.rs
@@ -26,7 +26,7 @@ use engine_executors::{
     transaction_registry::TransactionRegistry,
     webhook::WebhookJobHandler,
 };
-use twmq::{Queue, error::TwmqError};
+use twmq::{Queue, error::TwmqError, redis::aio::ConnectionManager};
 use vault_sdk::VaultClient;
 use vault_types::{
     RegexRule, Rule,
@@ -37,11 +37,12 @@ use vault_types::{
 use crate::chains::ThirdwebChainService;
 
 pub struct ExecutionRouter {
+    pub redis: ConnectionManager,
+    pub namespace: Option<String>,
     pub webhook_queue: Arc<Queue<WebhookJobHandler>>,
     pub external_bundler_send_queue: Arc<Queue<ExternalBundlerSendHandler<ThirdwebChainService>>>,
     pub userop_confirm_queue: Arc<Queue<UserOpConfirmationHandler<ThirdwebChainService>>>,
     pub eoa_executor_queue: Arc<Queue<EoaExecutorWorker<ThirdwebChainService>>>,
-    pub eoa_executor_store: Arc<EoaExecutorStore>,
     pub eip7702_send_queue: Arc<Queue<Eip7702SendHandler<ThirdwebChainService>>>,
     pub eip7702_confirm_queue: Arc<Queue<Eip7702ConfirmationHandler<ThirdwebChainService>>>,
     pub transaction_registry: Arc<TransactionRegistry>,
@@ -409,8 +410,15 @@ impl ExecutionRouter {
             transaction_type_data: transaction.transaction_type_data.clone(),
         };
 
+        let eoa_executor_store = EoaExecutorStore::new(
+            self.redis.clone(),
+            self.namespace.clone(),
+            eoa_execution_options.from,
+            base_execution_options.chain_id,
+        );
+
         // Add transaction to the store
-        self.eoa_executor_store
+        eoa_executor_store
             .add_transaction(eoa_transaction_request)
             .await
             .map_err(|e| TwmqError::Runtime {
